refactor(migration): gather the migration tooling under script/migrations
The migration tooling had grown to fifteen files sitting directly in
`script/`, interleaved with deploy helpers, devnet setup and coverage
scripts that have nothing to do with it. They now live together, and the
`migration` prefix that distinguished them goes away — inside the folder
it says nothing the path does not.

`script/migrate.ts` stays put as the operator entry point; the fixture
corpus keeps its own subfolder under the new one. The pre-migration
family stays in `script/`: it is a separate command with its own runbook,
and `premigrationIndex` and the TheGraph export belong beside it rather
than with the phases that consume them.

`extractFixture` derived its root by walking one level up from itself,
which the move broke. It now anchors on the directory that holds
`fixtures/`, so the next move cannot silently point it somewhere else.
