+++ contracts/package.json
@@ -63,8 +63,8 @@
     "coverage": "bun run coverage:forge && bun run coverage:hardhat && bun run coverage:reports",
     "clean": "forge clean && hardhat clean && rm -rf coverage/",
     "devnet": "bun run compile && bun ./script/runDevnet.ts",
-    "migration": "bun ./script/migration.ts",
-    "migration:fixture": "bun ./script/migrationFixture.ts",
+    "migration": "bun ./script/migrate.ts",
+    "migration:fixture": "bun ./script/migrations/fixture.ts",
     "docs:addresses": "bun ./script/generateAddressDocs.ts",
     "aakit": "bun ./script/deployAAkit.ts",
     "verify": "bun ./script/verify.ts",
