+++ contracts/plugins/migration/tasks/premigration-run.ts
@@ -1,8 +1,8 @@
 import type { NewTaskActionFunction } from "hardhat/types/tasks";
 
-import { parseMigrationNetwork } from "../../../script/migrationPlumbing.js";
+import { parseMigrationNetwork } from "../../../script/migrations/plumbing.js";
 
-import { runPreMigrationCommand } from "../../../script/migration.js";
+import { runPreMigrationCommand } from "../../../script/migrate.js";
 import {
   defaultHardhatSigner,
   nonEmptyString,
