+++ contracts/plugins/migration/tasks/set-v1-reverse-default-resolver.ts
@@ -1,8 +1,8 @@
 import type { NewTaskActionFunction } from "hardhat/types/tasks";
 
-import { parseMigrationNetwork } from "../../../script/migrationPlumbing.js";
+import { parseMigrationNetwork } from "../../../script/migrations/plumbing.js";
 
-import { setV1ReverseDefaultResolver } from "../../../script/migration.js";
+import { setV1ReverseDefaultResolver } from "../../../script/migrate.js";
 import {
   defaultHardhatPrivateKey,
   nonEmptyString,
