+++ contracts/plugins/migration/tasks/smoke-v2-registrar.ts
@@ -1,8 +1,8 @@
 import type { NewTaskActionFunction } from "hardhat/types/tasks";
 
-import { parseMigrationNetwork } from "../../../script/migrationPlumbing.js";
+import { parseMigrationNetwork } from "../../../script/migrations/plumbing.js";
 
-import { runV2RegistrarSmoke } from "../../../script/migration.js";
+import { runV2RegistrarSmoke } from "../../../script/migrate.js";
 import {
   addressForPrivateKey,
   defaultHardhatPrivateKey,
