+++ contracts/plugins/migration/tasks/utils.ts
@@ -1,7 +1,7 @@
 import { getAddress, type Address } from "viem";
 import { privateKeyToAccount } from "viem/accounts";
 
-import { isTenderlyVirtualRpc } from "../../../script/migration.js";
+import { isTenderlyVirtualRpc } from "../../../script/migrate.js";
 
 export { isTenderlyVirtualRpc };
